media: add usePrefersDark
