+++ packages/media/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@solid-primitives/media",
-  "version": "1.3.0",
+  "version": "2.0.0",
   "description": "Primitives for media query and device features",
   "author": "David Di Biase <dave.dibiase@gmail.com>",
   "contributors": [
@@ -22,7 +22,8 @@
     "stage": 3,
     "list": [
       "createMediaQuery",
-      "createBreakpoints"
+      "createBreakpoints",
+      "usePrefersDark"
     ],
     "category": "Display & Media"
   },
@@ -60,11 +61,11 @@
     "primitives"
   ],
   "dependencies": {
-    "@solid-primitives/utils": "^1.3.0",
-    "@solid-primitives/event-listener": "^2.0.0"
+    "@solid-primitives/utils": "^2.1.1",
+    "@solid-primitives/event-listener": "^2.2.0",
+    "@solid-primitives/rootless": "^1.1.0"
   },
   "devDependencies": {
-    "c8": "^7.11.0",
     "jsdom": "^19.0.0",
     "prettier": "^2.7.1",
     "solid-testing-library": "^0.2.0",
@@ -74,9 +75,10 @@
     "vitest": "^0.7.4",
     "unocss": "^0.39.1",
     "vite": "^2.9.12",
-    "vite-plugin-solid": "2.2.6"
+    "vite-plugin-solid": "2.2.6",
+    "solid-js": "^1.4.4"
   },
   "peerDependencies": {
-    "solid-js": "^1.3.1"
+    "solid-js": "^1.4.0"
   }
 }
