+++ packages/media/src/server.ts
@@ -4,7 +4,7 @@ import type * as API from "./index";
 
 export const makeMediaQueryListener: typeof API.makeMediaQueryListener = () => noop;
 
-export const createMediaQuery: typeof API.default =
+export const createMediaQuery: typeof API.createMediaQuery =
   (_, fallback = false) =>
   () =>
     fallback;
@@ -12,4 +12,7 @@ export const createMediaQuery: typeof API.default =
 export const createBreakpoints: typeof API.createBreakpoints = (breakpoints, options) =>
   options?.fallbackState || getEmptyMatchesFromBreakpoints(breakpoints);
 
-export default createMediaQuery;
+export const usePrefersDark: typeof API.usePrefersDark =
+  (fallback = false) =>
+  () =>
+    fallback;
